+++ tests/schedulers/scheduler_test.py
@@ -4,7 +4,7 @@ import pytest
 
 from ribs.archives import GridArchive
 from ribs.emitters import GaussianEmitter
-from ribs.schedulers import Scheduler
+from ribs.schedulers import BanditScheduler, Scheduler
 
 from ..archives.grid_archive_test import assert_archive_elite_batch
 
@@ -69,6 +69,63 @@ def test_ask_fails_when_called_twice(scheduler_fixture):
     with pytest.raises(RuntimeError):
         scheduler.ask()
         scheduler.ask()
+
+
+@pytest.mark.parametrize("archive_type", ["Scheduler", "BanditScheduler"])
+def test_warn_nothing_added_to_archive(archive_type):
+    archive = GridArchive(solution_dim=2,
+                          dims=[100, 100],
+                          ranges=[(-1, 1), (-1, 1)],
+                          threshold_min=1.0)
+    emitters = [GaussianEmitter(archive, sigma=1, x0=[0.0, 0.0], batch_size=4)]
+    if archive_type == "Scheduler":
+        scheduler = Scheduler(archive, emitters)
+    else:
+        scheduler = BanditScheduler(archive, emitters, 1)
+
+    _ = scheduler.ask()
+    with pytest.warns(UserWarning):
+        scheduler.tell(
+            # All objectives are below threshold_min of 1.0.
+            objective_batch=np.zeros(4),
+            # Arbitrary measures.
+            measures_batch=np.linspace(-1, 1, 4 * 2).reshape((4, 2)),
+        )
+
+
+@pytest.mark.parametrize("archive_type", ["Scheduler", "BanditScheduler"])
+def test_warn_nothing_added_to_result_archive(archive_type):
+    archive = GridArchive(solution_dim=2,
+                          dims=[100, 100],
+                          ranges=[(-1, 1), (-1, 1)],
+                          threshold_min=-np.inf)
+    result_archive = GridArchive(solution_dim=2,
+                                 dims=[100, 100],
+                                 ranges=[(-1, 1), (-1, 1)],
+                                 threshold_min=10.0)
+    emitters = [GaussianEmitter(archive, sigma=1, x0=[0.0, 0.0], batch_size=4)]
+    if archive_type == "Scheduler":
+        scheduler = Scheduler(
+            archive,
+            emitters,
+            result_archive=result_archive,
+        )
+    else:
+        scheduler = BanditScheduler(
+            archive,
+            emitters,
+            1,
+            result_archive=result_archive,
+        )
+
+    _ = scheduler.ask()
+    with pytest.warns(UserWarning):
+        scheduler.tell(
+            # All objectives are below threshold_min of 1.0.
+            objective_batch=np.zeros(4),
+            # Arbitrary measures.
+            measures_batch=np.linspace(-1, 1, 4 * 2).reshape((4, 2)),
+        )
 
 
 @pytest.mark.parametrize("tell_metadata", [True, False],