+++ HISTORY.md
@@ -1,11 +1,12 @@
 # History
 
-## 0.5.0 (Forthcoming)
+## 0.5.0
 
 ### Changelog
 
 #### API
 
+- Schedulers warn if no solutions are inserted into archive (#320)
 - Implement `BanditScheduler` (#299)
 - **Backwards-incompatible:** Implement Scalable CMA-ES Optimizers (#274, #288)
 - Make ribs.emitters.opt public (#281)
@@ -31,7 +32,7 @@
   - **Backwards-incompatible:** `behavior_dim` in archives is now `measure_dim`
   - Rename `n_solutions` to `batch_size` in `Scheduler`.
 - Add `GradientArborescenceEmitter`, which is used to implement CMA-MEGA (#240,
-  #263, #264, #282)
+  #263, #264, #282, #321)
 - Update emitter `tell()` docstrings to no longer say "Inserts entries into
   archive" (#247)
 - Expose `emitter.restarts` as a property (#248)
@@ -99,6 +100,8 @@
 
 #### Documentation
 
+- Speed up lunar lander tutorial (#319)
+- Add DQDTutorial (#267)
 - Remove examples extra in favor of individual example deps (#306)
 - Facilitate linking to latest version of documentation (#300)
 - Update lunar lander tutorial with v0.5.0 features (#292)